Black Litterman Model
The Black-Litterman asset allocation model, created by Fischer Black and Robert Litterman of Goldman, Sachs & Company, is a sophisticated method used to overcome the problem of unintuitive, highly-concentrated, input-sensitive portfolios. Input sensitivity is a well-documented problem with mean- variance optimization and is the most likely reason that more portfolio managers do not use the Markowitz paradigm, in which return is maximized for a given level of risk. The Black-Litterman Model uses a Bayesian approach to combine the subjective views of an investor regarding the expected returns of one or more assets with the market equilibrium vector (the prior distribution) of expected returns to form a new, mixed estimate of expected returns. The resulting new vector of returns (the posterior distribution) is described as a complex, weighted average of the investor’s views and the market equilibrium.

Change the parameters to your requirement.
Disclamer : I don't suggest you investing with help of this code analysis. The code is purely for education purpose and needs refinement before usage in real markets.